+++ package.json
@@ -11,15 +11,15 @@
     "license": "BSD-3-Clause",
     "repository": "git://github.com/hapijs/vise",
     "files": [
-        "lib",
+        "src",
         "README.md"
     ],
     "type": "module",
-    "types": "lib/index.d.ts",
+    "types": "src/index.d.mts",
     "exports": {
         ".": {
-            "types": "./lib/index.d.ts",
-            "default": "./lib/index.js"
+            "types": "./src/index.d.mts",
+            "default": "./src/index.mjs"
         }
     },
     "scripts": {
